requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openpyxl==3.1.2
python-multipart==0.0.9
python-dotenv==1.0.1
